+++ src/background/index.ts
@@ -1,12 +1,38 @@
 import browser from 'webextension-polyfill'
-
-// import { onMessage, sendMessage } from 'webext-bridge'
 import { setupAllMsgLstnrs } from './messageListeners'
 
 browser.runtime.onInstalled.addListener((): void => {
   // eslint-disable-next-line no-console
   console.log('Extension installed')
 })
+
+function isExtensionUri(url: string) {
+  return new URL(url).origin === new URL(browser.runtime.getURL('')).origin
+}
+
+// eslint-disable-next-line node/prefer-global/process
+if (process.env.FIREFOX) {
+  browser.webRequest.onBeforeSendHeaders.addListener(
+    (details) => {
+      const requestHeaders: browser.WebRequest.HttpHeaders = []
+      if (details.documentUrl) {
+        const url = new URL(details.documentUrl)
+        const extensionUri = isExtensionUri(details.documentUrl)
+        details.requestHeaders = details.requestHeaders || []
+        for (let i = 0; i < details.requestHeaders.length; i++) {
+          if (details.requestHeaders[i].name.toLowerCase() === 'origin' || details.requestHeaders[i].name.toLowerCase() === 'referer')
+            requestHeaders.push({ name: details.requestHeaders[i].name, value: extensionUri ? 'https://www.bilibili.com' : url.origin })
+          else
+            requestHeaders.push(details.requestHeaders[i])
+        }
+
+        return { ...details, requestHeaders }
+      }
+    },
+    { urls: ['<all_urls>'] },
+    ['blocking', 'requestHeaders'],
+  )
+}
 
 // Setup all message listeners
 setupAllMsgLstnrs()