+++ src/components/Settings/components/SearchPage.vue
@@ -1,44 +1,13 @@
 <script lang="ts" setup>
 import { settings } from '~/logic'
 
-const searchBarFocusCharacters = computed<{ name: string, url: string }[]>(() => {
-  return [
-    { name: '22 娘', url: 'https://pic.imgdb.cn/item/64d4f8891ddac507cc772ce5.png' },
-    { name: '33 娘', url: 'https://cdn.jsdelivr.net/gh/hakadao/bilibili-simple-home@master/img/searchBar_33_2.png' },
-    { name: '22 娘', url: 'https://pic.imgdb.cn/item/64d4fb391ddac507cc7ec5ae.png' },
-    { name: '33 娘', url: 'https://pic.imgdb.cn/item/64d4fd251ddac507cc8458fa.png' },
-  ]
-})
-const wallpapers = computed<Array<{ name: string, url: string, thumbnail: string }>>(() => {
-  return [
-    {
-      name: 'Unsplash Random Nature Image',
-      url: 'https://source.unsplash.com/1920x1080/?nature',
-      thumbnail: 'https://source.unsplash.com/1920x1080/?nature',
-    },
-    {
-      name: 'BML2019 VR (pid: 74271400)',
-      url: 'https://pic.imgdb.cn/item/638e1d63b1fccdcd36103811.jpg',
-      thumbnail: 'https://pic.imgdb.cn/item/64ac5e341ddac507cc750ae8.jpg',
-    },
-    {
-      name: '2020 拜年祭活动',
-      url: 'https://pic.imgdb.cn/item/638e1d7ab1fccdcd36106346.jpg',
-      thumbnail: 'https://pic.imgdb.cn/item/64ac5f251ddac507cc7658af.jpg',
-    },
-    {
-      name: '2020 BDF',
-      url: 'https://pic.imgdb.cn/item/63830f1816f2c2beb1868554.jpg',
-      thumbnail: 'https://pic.imgdb.cn/item/64ac5fc01ddac507cc77224e.jpg',
-    },
-  ]
-})
+const { searchBarCharacters, wallpapers, getBewlyImage } = useBewlyImage()
 
 watch(() => settings.value.individuallySetSearchPageWallpaper, (newValue) => {
   if (newValue)
-    document.documentElement.style.backgroundImage = `url(${settings.value.searchPageWallpaper})`
+    document.documentElement.style.backgroundImage = `url(${getBewlyImage(settings.value.searchPageWallpaper)})`
   else
-    document.documentElement.style.backgroundImage = `url(${settings.value.wallpaper})`
+    document.documentElement.style.backgroundImage = `url(${getBewlyImage(settings.value.wallpaper)})`
 })
 
 function changeSearchBarFocusCharacter(url: string) {
@@ -113,37 +82,20 @@ function changeWallpaper(url: string) {
           >
             <tabler:photo-off text="3xl $bew-text-3" />
           </picture>
-          <Tooltip v-for="item in searchBarFocusCharacters" :key="item.url" placement="top" :content="item.name" aspect-square>
+          <Tooltip v-for="item in searchBarCharacters" :key="item.url" placement="top" :content="item.name" aspect-square>
             <picture
               aspect-square bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
               un-border="4 transparent" w-full
               :class="{ 'selected-wallpaper': settings.searchPageSearchBarFocusCharacter === item.url }"
               @click="changeSearchBarFocusCharacter(item.url)"
             >
-              <img :src="item.url" alt="" w-full h-full object-contain>
+              <img
+                :src="getBewlyImage(item.url)" alt="" loading="lazy"
+                w-full h-full object-contain
+              >
             </picture>
           </Tooltip>
         </div>
-        <!-- <div flex items-center gap-4>
-          <div>
-            <picture
-              aspect-square bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
-              un-border="4 transparent" cursor-pointer shrink-0
-              w="xl:1/8 lg:1/6 md:1/5" p--4
-            >
-              <img
-                v-if="settings.searchPageSearchBarFocusCharacter" :src="settings.searchPageSearchBarFocusCharacter" alt="" w-full h-full object-contain
-                onerror="this.style.display='none'; this.onerror=null;"
-              >
-            </picture>
-          </div>
-          <div>
-            <Input v-model="settings.searchPageSearchBarFocusCharacter" w-full />
-            <p color="sm $bew-text-3" mt-2>
-              {{ $t('settings.image_url_hint') }}
-            </p>
-          </div>
-        </div> -->
       </SettingsItem>
     </SettingsItemGroup>
 
@@ -182,7 +134,7 @@ function changeWallpaper(url: string) {
         </SettingsItem>
 
         <SettingsItem v-if="settings.searchPageWallpaperMode === 'buildIn'" :title="$t('settings.choose_ur_wallpaper')" next-line>
-          <div grid="~ xl:cols-4 lg:cols-3 cols-2  gap-4">
+          <div grid="~ xl:cols-5 lg:cols-4 cols-3 gap-4">
             <picture
               aspect-video bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
               un-border="4 transparent" cursor-pointer
@@ -199,7 +151,7 @@ function changeWallpaper(url: string) {
                 :class="{ 'selected-wallpaper': settings.searchPageWallpaper === item.url }"
                 @click="changeWallpaper(item.url)"
               >
-                <img :src="item.thumbnail" alt="" w-full h-full object-cover>
+                <img :src="getBewlyImage(item.thumbnail)" alt="" w-full h-full object-cover>
               </picture>
             </Tooltip>
           </div>
@@ -209,10 +161,10 @@ function changeWallpaper(url: string) {
             <picture
               aspect-video bg="$bew-fill-1" rounded="$bew-radius" overflow-hidden
               un-border="4 transparent" cursor-pointer shrink-0
-              w="xl:1/4 lg:1/3 md:1/2"
+              w="xl:1/5 lg:1/4 md:1/3"
             >
               <img
-                v-if="settings.searchPageWallpaper" :src="settings.searchPageWallpaper" alt="" w-full h-full
+                v-if="settings.searchPageWallpaper" :src="getBewlyImage(settings.searchPageWallpaper)" alt="" w-full h-full
                 object-cover onerror="this.style.display='none'; this.onerror=null;"
               >
             </picture>