+++ src/components/TopBar/components/UserPanelPop.vue
@@ -4,13 +4,13 @@ import type { UserInfo, UserStat } from '../types'
 import { revokeAccessKey } from '~/utils/authProvider'
 import { getCSRF, getUserID, isHomePage } from '~/utils/main'
 import { numFormatter } from '~/utils/dataFormatter'
-import API from '~/background/msg.define'
 
 defineProps<{
   userInfo: UserInfo
 }>()
 
 const { t } = useI18n()
+const api = useApiClient()
 
 const mid = computed(() => {
   return getUserID()
@@ -22,6 +22,7 @@ const otherLinks = computed((): { name: string, url: string }[] => {
     { name: t('topbar.user_dropdown.uploads_manager'), url: 'https://member.bilibili.com/v2#/upload-manager/article' },
     { name: t('topbar.user_dropdown.b_coins_wallet'), url: 'https://pay.bilibili.com/' },
     { name: t('topbar.user_dropdown.orders'), url: 'https://show.bilibili.com/orderlist' },
+    { name: t('topbar.user_dropdown.workshop'), url: 'https://gf.bilibili.com?msource=main_station' },
     { name: t('topbar.user_dropdown.my_stream_info'), url: 'https://link.bilibili.com/p/center/index' },
     { name: t('topbar.user_dropdown.my_courses'), url: 'https://www.bilibili.com/cheese/mine/list' },
   ]
@@ -30,10 +31,7 @@ const otherLinks = computed((): { name: string, url: string }[] => {
 const userStat = reactive<UserStat>({} as UserStat)
 
 onMounted(() => {
-  browser.runtime
-    .sendMessage({
-      contentScriptQuery: API.USER.GET_USER_STAT,
-    })
+  api.user.getUserStat()
     .then((res) => {
       if (res.code === 0)
         Object.assign(userStat, res.data)
@@ -42,8 +40,7 @@ onMounted(() => {
 
 async function logout() {
   revokeAccessKey()
-  browser.runtime.sendMessage({
-    contentScriptQuery: API.AUTH.LOGOUT,
+  api.auth.logout({
     biliCSRF: getCSRF(),
   }).then(() => {
     location.reload()