+++ src/contentScripts/views/Home/components/Following.vue
@@ -2,7 +2,6 @@
 import type { Ref } from 'vue'
 import type { GridLayout } from '~/logic'
 import type { DataItem as MomentItem, MomentResult } from '~/models/moment/moment'
-import API from '~/background/msg.define'
 
 const props = defineProps<{
   gridLayout: GridLayout
@@ -20,6 +19,8 @@ const gridValue = computed((): string => {
     return '~ cols-1 xl:cols-2 gap-4'
   return '~ cols-1 gap-4'
 })
+
+const api = useApiClient()
 
 const videoList = reactive<MomentItem[]>([])
 const isLoading = ref<boolean>(false)
@@ -82,8 +83,7 @@ async function getFollowedUsersVideos() {
   emit('beforeLoading')
   isLoading.value = true
   try {
-    const response: MomentResult = await browser.runtime.sendMessage({
-      contentScriptQuery: API.MOMENT.GET_MOMENTS,
+    const response: MomentResult = await api.moment.getMoments({
       type: 'video',
       offset: offset.value,
       update_baseline: updateBaseline.value,